// 17-managing-query-keys/src/types/book.ts
export interface BookT {
  id: string;
  title: string;
  authors: string[];
  publisher: string;
  publishedDate: string; // ISO date string (YYYY-MM-DD)
  description: string; // HTML formatted string
  thumbnail: string; // URL string
  previewLink: string; // URL string
  averageRating: number;
  availableCopies: number;
  isCheckedOutByUser: boolean;
}

export interface SearchResponse {
  books: BookT[];
  currentPage: number;
  totalPages: number;
};